Create Date_three_forms.cpp
This is a program that reads three integer values representing the four-digit month, day, and year. The program display the date in one of the following formats, depending on the user's selection. If the user enters 1, the month will be displayed as a word (“January”), the day as an integer (11) and the year (1999). Entering 2 will display the month as a three  letter abbreviation (“Jan”), the day as an integer (11), and the year (1999). If entering 3, the month will be displayed as an integer (01), the day as an integer (11), and the year (1999).