Create Multiplication_tables.cpp
This program shows the multiplication tables. The programs must shows a menu with the tables, the user must choose which table he wants to be printed on the screen, the user can also choose up to what number the table will be displayed, for example the table of 5 up to number 20. At the end of displaying the table, the user can choose whether to return to the menu or end the program. If, while or do while and for are used.