Create Array_proyect.cpp
Program to save the notes of N students, N has to be greater than or equal to 20 or less than or equal to 30, this must be verified in the program, and in case the number of students is outside the range, it must be show the option to re-enter the number of students. Once we have the grades stored in an array, we have to add all the grades, show the class average, show the student with the highest grade, show students with the average grade (range from average –1 to average +1) and the student with the lowest grade. Students who have a grade less than or equal to 6 were assigned a job, this job adds one more point to their grade. The arrangement with the original notes must be used to generate another arrangement with the modified notes of these students. Print the two notes in different columns. At the end, the option to recalculate more student grades, or end the program, should be shown.