Create Alphabet.cpp
This is a program where the user can enter the alphabet in a string (in lowercase), then pass the data to another string, and then pass the alphabet of the second string to uppercase. Print everything on screen. (In the end we will have two strings, one with the alphabet in lowercase and one in uppercase).