Create Movies.cpp
This program saves twenty movies in a matrix  (movie name and year of release). It has a menu with three options: a) Shows movies from oldest to newest. b) Shows movies from newest to oldest. c) Shows movies from a specific year. Also, it has the option to return to menu or exit the program.